requirements.txt
numpy
scipy
SoundFile
munch
pydub
pyyaml
librosa
matplotlib
pandas
click
tensorboard
openai-whisper
transformers>=4.30.0,<5.0.0
accelerate>=0.20.0
einops
einops-exts
tqdm
typing-extensions
a_unet
unidic-lite
fugashi
cutlet
xlstm
nltk
wandb
git+https://github.com/resemble-ai/monotonic_align.git
gradio>=4.0.0
plotly
